pyproject.toml
[project]
name = "gos-sanity"
version = "0.1.0"
description = "Sanity check: does SkillsBench reward respond to bundle composition?"
requires-python = ">=3.10"
dependencies = [
    "numpy>=1.26",
    "scipy>=1.12",
    "pandas>=2.2",
    "pyyaml>=6.0",
    "tqdm>=4.66",
    "sentence-transformers>=2.7",   # local embeddings for unrelated-swap candidate filtering
    "openai>=1.40",                 # primary agent backend per preregistration
    "anthropic>=0.34",              # secondary backend
]

[project.optional-dependencies]
local = ["vllm>=0.5"]               # for A100-hosted agent if API path is unavailable

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]
